+++ integration_tests/stacks/lambda-function-stack-legacy-datadog-api.ts
@@ -7,33 +7,32 @@
  */
 
 import * as lambda from "aws-cdk-lib/aws-lambda";
-import { PythonFunction } from "@aws-cdk/aws-lambda-python-alpha";
+import { Stack, StackProps, App } from "aws-cdk-lib";
 import { LambdaRestApi, LogGroupLogDestination } from "aws-cdk-lib/aws-apigateway";
 import { LogGroup } from "aws-cdk-lib/aws-logs";
-import { Stack, StackProps, App } from "aws-cdk-lib";
-import { DatadogLambda } from "../../src/index";
+// use the legacy Datadog class instead of the new DatadogLambda
+import { Datadog } from "../../src/index";
 
 export class ExampleStack extends Stack {
   constructor(scope: App, id: string, props?: StackProps) {
     super(scope, id, props);
 
-    const lambdaPythonFunction = new PythonFunction(this, "HelloHandler", {
-      runtime: lambda.Runtime.PYTHON_3_7,
-      entry: __dirname + "/../../../lambda",
-      index: "example-python.py",
-      handler: "handler",
+    const lambdaFunction = new lambda.Function(this, "HelloHandler", {
+      runtime: lambda.Runtime.NODEJS_14_X,
+      code: lambda.Code.fromInline("test"),
+      handler: "lambdaFunction.handler",
     });
 
     const restLogGroup = new LogGroup(this, "restLogGroup");
     new LambdaRestApi(this, "rest-test", {
-      handler: lambdaPythonFunction,
+      handler: lambdaFunction,
       deployOptions: {
         accessLogDestination: new LogGroupLogDestination(restLogGroup),
       },
     });
 
-    const datadogLambda = new DatadogLambda(this, "Datadog", {
-      pythonLayerVersion: 46,
+    const datadog = new Datadog(this, "Datadog", {
+      nodeLayerVersion: 62,
       extensionLayerVersion: 10,
       enableDatadogTracing: true,
       flushMetricsToLogs: true,
@@ -41,13 +40,13 @@ export class ExampleStack extends Stack {
       apiKey: "1234",
       site: "datadoghq.com",
     });
-    datadogLambda.addLambdaFunctions([lambdaPythonFunction]);
-    datadogLambda.addForwarderToNonLambdaLogGroups([restLogGroup]);
+    datadog.addLambdaFunctions([lambdaFunction]);
+    datadog.addForwarderToNonLambdaLogGroups([restLogGroup]);
   }
 }
 
 const app = new App();
 const env = { account: "601427279990", region: "sa-east-1" };
-const stack = new ExampleStack(app, "lambda-python-function-stack", { env: env });
+const stack = new ExampleStack(app, "lambda-function-stack", { env: env });
 console.log("Stack name: " + stack.stackName);
 app.synth();
