+++ integration_tests/stacks/lambda-provided-arm-stack.ts
@@ -3,7 +3,7 @@ import * as s3 from "aws-cdk-lib/aws-s3";
 import { LambdaRestApi, LogGroupLogDestination } from "aws-cdk-lib/aws-apigateway";
 import { LogGroup } from "aws-cdk-lib/aws-logs";
 import { Stack, StackProps, App } from "aws-cdk-lib";
-import { Datadog } from "../../src/index";
+import { DatadogLambda } from "../../src/index";
 
 export class ExampleStack extends Stack {
   constructor(scope: App, id: string, props?: StackProps) {
@@ -31,14 +31,14 @@ export class ExampleStack extends Stack {
       },
     });
 
-    const datadogCDK = new Datadog(this, "Datadog", {
+    const datadogLambda = new DatadogLambda(this, "Datadog", {
       extensionLayerVersion: 49,
       apiKey: "1234",
       site: "datadoghq.com",
       sourceCodeIntegration: false,
     });
-    datadogCDK.addLambdaFunctions([providedLambda]);
-    datadogCDK.addForwarderToNonLambdaLogGroups([restLogGroup]);
+    datadogLambda.addLambdaFunctions([providedLambda]);
+    datadogLambda.addForwarderToNonLambdaLogGroups([restLogGroup]);
   }
 }
